fix(capture): wlr_buffer stores int width/height, drop the casts
The upstream wlroots wlr_buffer struct defines width/height as int, so
comparing them with QSize values needs no casts at all; the previous
uint32_t casts were based on a wrong assumption and reintroduced
sign-compare failures on the other operand order.

+++ waylib/src/server/utils/wextimagecapturesourcev1impl.cpp
@@ -361,8 +361,8 @@ void WExtImageCaptureSourceV1Impl::announceFrame()
     auto *tp = m_renderer ? m_renderer->wTextureProvider() : nullptr;
     if (!tp || !tp->wlrBuffer())
         return;
-    if (tp->wlrBuffer()->width != static_cast<uint32_t>(pixelSize.width())
-        || tp->wlrBuffer()->height != static_cast<uint32_t>(pixelSize.height())) {
+    if (tp->wlrBuffer()->width != pixelSize.width()
+        || tp->wlrBuffer()->height != pixelSize.height()) {
         return;
     }
